+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [Unreleased]
 
 - Allow inheritance from packet classes
+- Validate that `ConnectPacket` / `ConnAckPacket` have non-zero `receive_max`
 
 ## [0.7.0] - 2026-04-23